requirements.txt
pandas
numpy
python-dotenv
oracledb
psycopg2-binary
streamlit
plotly
xlsxwriter